Working 6 group max config
This gives a 6 group max config, Will make it more efficiant with a
PriorityQueue implementation of the Array. If tom's code wasn't so
abstract to begin with then none of this would be possible. He either
meant to make it abstract or he typed whatever came to mind as the
easiest way after 1 second of thinking...